Move common steps together
