+++ builder/amazon/chroot/step_attach_volume_test.go
@@ -1,11 +1,15 @@
 package chroot
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/hashicorp/packer/common/chroot"
+)
 
 func TestAttachVolumeCleanupFunc_ImplementsCleanupFunc(t *testing.T) {
 	var raw interface{}
 	raw = new(StepAttachVolume)
-	if _, ok := raw.(Cleanup); !ok {
+	if _, ok := raw.(chroot.Cleanup); !ok {
 		t.Fatalf("cleanup func should be a CleanupFunc")
 	}
 }
