+++ builder/amazon/chroot/step_early_unflock.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/hashicorp/packer/common/chroot"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 )
@@ -13,7 +14,7 @@ import (
 type StepEarlyUnflock struct{}
 
 func (s *StepEarlyUnflock) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
-	cleanup := state.Get("flock_cleanup").(Cleanup)
+	cleanup := state.Get("flock_cleanup").(chroot.Cleanup)
 	ui := state.Get("ui").(packer.Ui)
 
 	log.Println("Unlocking file lock...")
