+++ builder/amazon/chroot/step_mount_device_test.go
@@ -1,11 +1,15 @@
 package chroot
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/hashicorp/packer/common/chroot"
+)
 
 func TestMountDeviceCleanupFunc_ImplementsCleanupFunc(t *testing.T) {
 	var raw interface{}
 	raw = new(StepMountDevice)
-	if _, ok := raw.(Cleanup); !ok {
+	if _, ok := raw.(chroot.Cleanup); !ok {
 		t.Fatalf("cleanup func should be a CleanupFunc")
 	}
 }
