+++ common/chroot/communicator_test.go
