+++ common/chroot/interpolate_context_provider.go
@@ -1,0 +1,7 @@
+package chroot
+
+import "github.com/hashicorp/packer/template/interpolate"
+
+type interpolateContextProvider interface {
+	GetContext() interpolate.Context
+}
