+++ common/chroot/step_copy_files.go
