+++ common/chroot/step_copy_files_test.go
