+++ common/chroot/step_early_cleanup.go
